+++ src/hooks/useCalendarRange.ts
@@ -40,9 +40,18 @@ export const useCalendarRange = (
   /**
    * Middle date within the range to show current date in Calendar
    */
-  const middleDate = useMemo(() => {
+  const middleDate = useMemo((): Date => {
+    /**
+     * Return start of period date for agenda due to different calculation
+     * From: date
+     * To: date + days count
+     */
+    if (view === View.AGENDA) {
+      return dayjs(calendarTo).startOf('month').toDate();
+    }
+
     return new Date((calendarFrom.valueOf() + calendarTo.valueOf()) / 2);
-  }, [calendarFrom, calendarTo]);
+  }, [calendarFrom, calendarTo, view]);
 
   /**
    * Change Calendar View
@@ -80,6 +89,12 @@ export const useCalendarRange = (
       let view: View = action === Navigate.DATE ? View.DAY : currentView;
 
       /**
+       * Prevent incorrect navigation for Agenda view
+       * Add 1 day because agenda view has month days - 1 to keep end date in the last day of month
+       */
+      const newCurrentDate = view === View.AGENDA ? dayjs(newDate).add(1, 'day').toDate() : newDate;
+
+      /**
        * Open Week by clicking on day in Year View
        */
       if (currentView === View.YEAR && action === Navigate.DATE) {
@@ -88,8 +103,8 @@ export const useCalendarRange = (
 
       const unitType = getUnitType(view);
       const { from, to } = timeRange;
-      const newFrom = dayjs(newDate).startOf(unitType);
-      const newTo = dayjs(newDate).endOf(unitType);
+      const newFrom = dayjs(newCurrentDate).startOf(unitType);
+      const newTo = dayjs(newCurrentDate).endOf(unitType);
 
       /**
        * Change time range if one of dates are out of the current range